+++ LawnCare/frontend/app/src/main/res/values/strings.xml
@@ -1,4 +1,7 @@
 <resources>
     <string name="app_name">LawnCare</string>
     <string name="action_settings">Settings</string>
+    <string name="tab_contact">Contact</string>
+    <string name="tab_FAQ">FAQ</string>
+    <string name="tab_book">Book Now</string>
 </resources>
