+++ LawnCare/frontend/gradle.properties
@@ -22,3 +22,6 @@ org.gradle.jvmargs=-Xmx1536m
 android.useAndroidX=true
 # Automatically convert third-party libraries to use AndroidX
 android.enableJetifier=true
+API_TOKEN = "8c586758194612ffa97b9813dff58b99"
+API_KEY = "8b8888512eeae39d9e0c556d21facde6"
+API_SSL = "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"
